refactor: Renamed the directory of the JourneyLegsSummary-component
