+++ src/components/journey-legs-summary/JourneyLegsSummary.tsx
