+++ src/components/journey-legs-summary/index.tsx
